pyproject.toml
[tool.poetry]
name = "netbox-plugin-config-officer-2"
version = "0.1.6"
description = "NetBox plugin for collecting running configs, tracking diffs and checking compliance with templates"
authors = [
    "Michal Juskiewicz <m.juskiewicz66@gmail.com>",
    "Sergei Artemov <artemov.sergey1989@gmail.com>",
]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://github.com/Juzekkk/netbox-plugin-config-officer-2"
repository = "https://github.com/Juzekkk/netbox-plugin-config-officer-2"
packages = [
    { include = "config_officer" },
]
include = [
    "config_officer/templates/**/*",
]

[tool.poetry.dependencies]
python = "^3.10"
scrapli = ">=2024.1.30,<2026.0"
GitPython = ">=3.1,<4.0"
gitdb = ">=4.0.1,<5.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-django = "^4.8"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"